Added public interfaces:  ITag, and ITagFile
Design Change:  This change was implemented to avoid inheritance patterns using abstract classes. Because the reading component was not separated out from the Tag (frontend) class, there was no way to 1) open/read a tag, and 2) Load its properties into your inherited class for some of the parts of the spec:  1) Properties, and 2) File corruption reasons. The Tag object would work together with a new interface - created by the user. However, it would be best that the base inheritance is kept to the taglib library. Please consider taking this change as an addition to the current library.

No other changes were made.