Implemented ITag for the abstract Tag class. Unit tests are running successfully.
Design Impact:  The property ITag.IsCompilation was not used uniformly (for all tag types); but it was situated at the base of all the tags  (abstract class Tag). So, there was a need to put the public property into tags that wouldn't otherwise use it. However, this did not impact the functioning of the library.

Suggestions:  1) Separate base tags into groups that have common properties.  2) Remove the read/write code from the public properties; and put it into separate components. We need these methods to be where we can use them.

So, the design I've recommended is based on compontization by means of interfaces. There will be one more change set - separating out the read/write code.